Add fast abs w/tests
The optimized version takes around 65% of the cycles that the
C version does. Included are a consistency test (which tests
every return value from this abs with the return value from a
known good one), and a benchmark (which uses the BGB macros for
accurate clock cycle counting).

As this passes both the tests I think it's safe to be used by
default.

Credit to @aaaaaa123456789 for golfing the optimized version
down to 15 bytes from my initial attempt which was 20.