Clarify the use of lcc
Historically, GBDK was shipped with the modified LCC described in the
book "A Retargetable C Compiler: Design and Implementation." However,
SDCC has been the actual C compiler for GBDK, and the program "lcc" is
now the wrapper of SDCC. That is, "lcc" is the compiler driver rather
a C compiler for the sake of compatibility. This patch removed
misleading descriptions about LCC.

A compiler driver is a program which invokes the other components of
the tool set to process a program. The compiler driver uses command
line arguments and source file extensions to determine which compiler
or assembler to invoke for each source file, then sequence the
resulting output through the subsequent linker and conversion
utilities, relieving the user of the burden of invoking each of these
tools individually.