Add -keep_duplicate_tiles as option to png2asset
The option -keep_duplicate_tiles can be used to have it export all tiles, without removing duplicates.
This could for example be used together with -map and -tiles_only, to get a list of all tiles in your image. This can then be used with draw_image() to display it.