Fix gbdk-lib-install-ports target dependencies
The gbdk-lib-install-ports has an unwritten dependency on the
gbdk-lib-build target. the gbdk-lib-install-ports target fails if built
before the gbdk-lib-build target. In a non-parallelized build, these two
targets are built in the right order and the build works fine. In a
parallelized build, a race condition causes the targets to build in the
wrong order.